+++ data-raw/internal_data.R
@@ -2,5 +2,5 @@ source("data-raw/field_names.R")
 source("data-raw/updatesql.R")
 
 
-usethis::use_data(sql_tables, sql_versions, internal = TRUE, overwrite = TRUE)
+usethis::use_data(sql_tables, sql_fields, sql_versions, internal = TRUE, overwrite = TRUE)